feat(perps): prepare migration to perpscontroller to core
