+++ app/components/UI/Perps/controllers/perpsErrorCodes.ts
@@ -1,7 +1,7 @@
 /**
  * Error codes for PerpsController
  * These codes are returned to the UI layer for translation
- * Extracted to separate file to avoid circular dependencies with perpsErrorHandler
+ * Extracted to separate file to avoid circular dependencies with translatePerpsError
  */
 export const PERPS_ERROR_CODES = {
   CLIENT_NOT_INITIALIZED: 'CLIENT_NOT_INITIALIZED',
@@ -20,6 +20,29 @@ export const PERPS_ERROR_CODES = {
   IOC_CANCEL: 'IOC_CANCEL', // Order could not immediately match (insufficient liquidity)
   // Connection errors
   CONNECTION_TIMEOUT: 'CONNECTION_TIMEOUT',
+  // Validation errors - withdraw
+  WITHDRAW_ASSET_ID_REQUIRED: 'WITHDRAW_ASSET_ID_REQUIRED',
+  WITHDRAW_AMOUNT_REQUIRED: 'WITHDRAW_AMOUNT_REQUIRED',
+  WITHDRAW_AMOUNT_POSITIVE: 'WITHDRAW_AMOUNT_POSITIVE',
+  WITHDRAW_INVALID_DESTINATION: 'WITHDRAW_INVALID_DESTINATION',
+  WITHDRAW_ASSET_NOT_SUPPORTED: 'WITHDRAW_ASSET_NOT_SUPPORTED',
+  WITHDRAW_INSUFFICIENT_BALANCE: 'WITHDRAW_INSUFFICIENT_BALANCE',
+  // Validation errors - deposit
+  DEPOSIT_ASSET_ID_REQUIRED: 'DEPOSIT_ASSET_ID_REQUIRED',
+  DEPOSIT_AMOUNT_REQUIRED: 'DEPOSIT_AMOUNT_REQUIRED',
+  DEPOSIT_AMOUNT_POSITIVE: 'DEPOSIT_AMOUNT_POSITIVE',
+  DEPOSIT_MINIMUM_AMOUNT: 'DEPOSIT_MINIMUM_AMOUNT',
+  // Validation errors - order
+  ORDER_COIN_REQUIRED: 'ORDER_COIN_REQUIRED',
+  ORDER_LIMIT_PRICE_REQUIRED: 'ORDER_LIMIT_PRICE_REQUIRED',
+  ORDER_PRICE_POSITIVE: 'ORDER_PRICE_POSITIVE',
+  ORDER_UNKNOWN_COIN: 'ORDER_UNKNOWN_COIN',
+  ORDER_SIZE_POSITIVE: 'ORDER_SIZE_POSITIVE',
+  ORDER_PRICE_REQUIRED: 'ORDER_PRICE_REQUIRED',
+  ORDER_SIZE_MIN: 'ORDER_SIZE_MIN',
+  ORDER_LEVERAGE_INVALID: 'ORDER_LEVERAGE_INVALID',
+  ORDER_LEVERAGE_BELOW_POSITION: 'ORDER_LEVERAGE_BELOW_POSITION',
+  ORDER_MAX_VALUE_EXCEEDED: 'ORDER_MAX_VALUE_EXCEEDED',
 } as const;
 
 export type PerpsErrorCode =
