+++ app/components/UI/Perps/controllers/services/MarketDataService.ts
@@ -1,5 +1,5 @@
 import Logger from '../../../../../util/Logger';
-import { ensureError } from '../../utils/perpsErrorHandler';
+import { ensureError } from '../../../../../util/errorUtils';
 import {
   trace,
   endTrace,
