+++ app/components/UI/Perps/hooks/stream/usePerpsLiveCandles.ts
@@ -5,7 +5,7 @@ import { CandlePeriod, TimeDuration } from '../../constants/chartConfig';
 import { PERPS_CONSTANTS } from '../../constants/perpsConfig';
 import DevLogger from '../../../../../core/SDKConnect/utils/DevLogger';
 import Logger from '../../../../../util/Logger';
-import { ensureError } from '../../utils/perpsErrorHandler';
+import { ensureError } from '../../../../../util/errorUtils';
 
 // Stable empty candle data reference to prevent re-renders
 const EMPTY_CANDLE_DATA: CandleData = {
