+++ app/components/UI/Perps/providers/PerpsConnectionProvider.tsx
@@ -13,7 +13,7 @@ import { isE2E } from '../../../../util/test/utils';
 import PerpsConnectionErrorView from '../components/PerpsConnectionErrorView';
 import type { ReconnectOptions } from '../types/perps-types';
 import Logger from '../../../../util/Logger';
-import { ensureError } from '../utils/perpsErrorHandler';
+import { ensureError } from '../../../../util/errorUtils';
 import { PERPS_CONSTANTS } from '../constants/perpsConfig';
 
 export interface PerpsConnectionContextValue {
