+++ app/components/UI/Perps/utils/hyperLiquidValidation.test.ts
@@ -16,6 +16,7 @@ import {
 } from './hyperLiquidValidation';
 import type { CaipAssetId, Hex } from '@metamask/utils';
 import type { GetSupportedPathsParams } from '../controllers/types';
+import { PERPS_ERROR_CODES } from '../controllers/perpsErrorCodes';
 
 jest.mock('@metamask/utils', () => ({
   isValidHexAddress: (address: string) => /^0x[0-9a-fA-F]{40}$/.test(address),
@@ -92,7 +93,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         success: false,
-        error: 'Unknown error occurred',
+        error: PERPS_ERROR_CODES.UNKNOWN_ERROR,
         data: null,
       });
     });
@@ -136,8 +137,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'assetId is required for withdrawals. Please provide an asset ID in CAIP format (e.g., eip155:42161/erc20:0xaf88d065e77c8cC2239327C5EDb3A432268e5831)',
+        error: PERPS_ERROR_CODES.WITHDRAW_ASSET_ID_REQUIRED,
       });
     });
 
@@ -152,8 +152,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount must be a positive number. Amount must be a positive number (received: 0)',
+        error: PERPS_ERROR_CODES.WITHDRAW_AMOUNT_POSITIVE,
       });
     });
 
@@ -169,8 +168,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Invalid destination address format: invalid-address. Address must be a valid Ethereum address starting with 0x',
+        error: PERPS_ERROR_CODES.WITHDRAW_INVALID_DESTINATION,
       });
     });
 
@@ -184,8 +182,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'amount is required for withdrawals. Please specify the amount to withdraw',
+        error: PERPS_ERROR_CODES.WITHDRAW_AMOUNT_REQUIRED,
       });
     });
 
@@ -200,8 +197,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount must be a positive number. Amount must be a positive number (received: -10)',
+        error: PERPS_ERROR_CODES.WITHDRAW_AMOUNT_POSITIVE,
       });
     });
 
@@ -255,8 +251,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'AssetId is required for deposit validation. Please provide an asset ID in CAIP format (e.g., eip155:42161/erc20:0xaf88d065e77c8cC2239327C5EDb3A432268e5831)',
+        error: PERPS_ERROR_CODES.DEPOSIT_ASSET_ID_REQUIRED,
       });
     });
 
@@ -271,8 +266,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount is required and must be greater than 0. Please specify the amount to deposit',
+        error: PERPS_ERROR_CODES.DEPOSIT_AMOUNT_REQUIRED,
       });
     });
 
@@ -288,8 +282,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount must be a positive number. Amount must be a positive number (received: 0)',
+        error: PERPS_ERROR_CODES.DEPOSIT_AMOUNT_POSITIVE,
       });
     });
 
@@ -305,8 +298,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount must be a positive number. Amount must be a positive number (received: -10)',
+        error: PERPS_ERROR_CODES.DEPOSIT_AMOUNT_POSITIVE,
       });
     });
 
@@ -322,8 +314,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Minimum deposit amount is 5 USDC. Current amount: 4.99, required minimum: 5',
+        error: PERPS_ERROR_CODES.DEPOSIT_MINIMUM_AMOUNT,
       });
     });
 
@@ -339,8 +330,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Minimum deposit amount is 11 USDC. Current amount: 10.99, required minimum: 11',
+        error: PERPS_ERROR_CODES.DEPOSIT_MINIMUM_AMOUNT,
       });
     });
 
@@ -381,8 +371,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Minimum deposit amount is 5 USDC. Current amount: 4, required minimum: 5',
+        error: PERPS_ERROR_CODES.DEPOSIT_MINIMUM_AMOUNT,
       });
     });
 
@@ -398,8 +387,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Amount must be a positive number. Amount must be a positive number (received: invalid)',
+        error: PERPS_ERROR_CODES.DEPOSIT_AMOUNT_POSITIVE,
       });
     });
   });
@@ -437,10 +425,7 @@ describe('hyperLiquidValidation', () => {
       const result = validateAssetSupport(assetId, supportedRoutes);
 
       expect(result.isValid).toBe(false);
-      expect(result.error).toContain('is not supported for withdrawals');
-      expect(result.error).toContain(
-        '0xaf88d065e77c8cc2239327c5edb3a432268e5831',
-      );
+      expect(result.error).toBe(PERPS_ERROR_CODES.WITHDRAW_ASSET_NOT_SUPPORTED);
     });
 
     it('should handle empty supported routes', () => {
@@ -451,8 +436,7 @@ describe('hyperLiquidValidation', () => {
       const result = validateAssetSupport(assetId, supportedRoutes);
 
       expect(result.isValid).toBe(false);
-      expect(result.error).toContain('is not supported for withdrawals');
-      expect(result.error).toContain('Supported assets: ');
+      expect(result.error).toBe(PERPS_ERROR_CODES.WITHDRAW_ASSET_NOT_SUPPORTED);
     });
 
     it('should support case-insensitive asset matching for contract addresses', () => {
@@ -499,8 +483,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Insufficient balance. Available: 100, Requested: 150. You need 50.000000 more to complete this withdrawal',
+        error: PERPS_ERROR_CODES.WITHDRAW_INSUFFICIENT_BALANCE,
       });
     });
 
@@ -515,8 +498,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error:
-          'Insufficient balance. Available: 0, Requested: 10. You need 10.000000 more to complete this withdrawal',
+        error: PERPS_ERROR_CODES.WITHDRAW_INSUFFICIENT_BALANCE,
       });
     });
   });
@@ -678,7 +660,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Coin is required for orders',
+        error: PERPS_ERROR_CODES.ORDER_COIN_REQUIRED,
       });
     });
 
@@ -693,7 +675,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Price must be a positive number if provided',
+        error: PERPS_ERROR_CODES.ORDER_PRICE_POSITIVE,
       });
     });
 
@@ -731,7 +713,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Price is required for limit orders',
+        error: PERPS_ERROR_CODES.ORDER_LIMIT_PRICE_REQUIRED,
       });
     });
   });
@@ -758,7 +740,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Unknown coin: UNKNOWN',
+        error: PERPS_ERROR_CODES.ORDER_UNKNOWN_COIN,
       });
     });
 
@@ -768,7 +750,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Unknown coin: BTC',
+        error: PERPS_ERROR_CODES.ORDER_UNKNOWN_COIN,
       });
     });
 
@@ -777,7 +759,7 @@ describe('hyperLiquidValidation', () => {
 
       expect(result).toEqual({
         isValid: false,
-        error: 'Unknown coin: btc',
+        error: PERPS_ERROR_CODES.ORDER_UNKNOWN_COIN,
       });
     });
   });
